Additional architure cleanup
Cleaned up calls to CharacterController2D.Move (we might mod it to get those problems out of the way) and changed which SerializableDictionary we're to decrease asset clutter by using a unity package instead.
Also, ActorAbility is an abstract class now, because Interfaces can't be serialized. So, ActorAbility inherits from  Monobehavior and ActorAbilityFunction<,> inherits from ActorAbility. Same functionality, but with a few extra hoops.