Player Actor Derivatives and new Movement Controller scripts
Switched more properties from private to protected

Also got rid of the Brackeys CharacterController2D because it had too much fluff that made it hard to use, it's been replaced with a very simple custom controller.

Other than that, this commit is just copying old player code over from the prototype and translating it into the new actor architecture.