Ability Register
The AbilityRegister is a publically accessible, type-strict register of keys being used to refer to certain abilities in abilityDicts, which exist in various ActorAbilityInitiators. Basically, it's a big list of keys everyone can know instead of actually knowing the keys, and the compiler will yell at you if you get it wrong!

The Player and Gluttony will most likely need to be refactored a bit to use this new registry, tasks for doing so will be added soon.