Pausing development on PrideSin
PrideSin is currently in development. However, Design (Roy) has requested more time to finish the design of Pride's sin. So, PrideSin.cs has been moved into .PausedPrideDevelopment, and will remain there until Design (Roy) is finished.

Files that begin with . are ignored by Unity. .PausedPrideDevelopment is in the project, but cannot be viewed in Editor nor will any files inside of it be compiled. Simply rename the folder to return PrideSin to the project.