Pride Ability: Shockwave
Added a shockwave projectile ability, designed to be used by Pride.

PrideShockwave is the Ability. It launches...
PrideWaveProjectile is an BasicHitbox that moves and grows with time. It mirrors the cone-like behavior of the wave while also moving in a direction.

Also created a few helper things:
Renamed msic to Misc
BasicHitbox is a box that just hurts anything that enters its trigger collider and that's it
Rotation2D is a static class that holds rotation-related helper functions, mainly because most rotation functions in Unity don't map correctly to 2D spaces.

I won't lie, I (Thomas) am mega stoopidbrain mode. This might be missing some stoff. But that's what testing is for!