Finished Pride Sin
The design for Pride's Sin has been finished. Its effect has been cut, and the statue was already done, so no significant changes were needed other than putting the prefab together.

Most of the file changes you see are from the addition of Text Mesh Pro to the project. It does text as a vector mesh instead of a bitmap. This prevets very large or very small text from being blurry, which is important because the UI prompt on the sin statues is (relatively) very small.