Upgraded ActorAbility to take object arrays to act as InternInvoke arguments
Invoke can take a bunch of objects (as a params array) and then decide how to translate it into an InvokeParam array to give to InternInvoke.

Comes with a method that converts the array of objects to InvokeParam one-for-one and public boolean to toggle if the ability should activate even if the object params aren't good.

Check ActorAbility's documentation for more detail