Pride Ability: Phase 2 Special
Pride uses the exact same Charge for his Phase 2 Special, except it doesn't knockback. Instead, it reduces the player's health to 1 (or kills them if they're at 1 or less.

Since the knockback is actually just pride's flex punch, I decided to do the same thing for Phase 2: make a seperate ability that gets invoked when the charge ends.