More Statue Bugfixes
Specifically, changed how damage is dealt to someone with damage resistance in ActorHealth from a division to an interpolation. This requires ActorHealth.takeDamage to take an additional, optional argument to specify if damage should be bypassed, but now there are no edge cases where the mathematics of damage resistance cause certain values to produce unintended results (specifically dividing by 0 to bypass 100% damage resistance).