Sloth Startup Crash Fix
Fixed a bug where Sloth would crash when its scene was loaded due to trying to start dialogue with the player before the player had finished loading.

Also included some QoL features that should have been there in the first place (UI and good Boss Camera)

Introduced a different bug where clicking to advance Sloth' dialogue freezes the dialogue. However, pressing ESC to skip the dialogue starts the fight as normal, so it's not as big of a deal.