Added TrySetTrigger to ActorAnimationController
It basically checks to see if a certain trigger exists and then calls it if it does, in fact, exist. It's pretty slow, so try to avoid using it constantly.

This was created mainly as a way to trigger animations without needing type conversions on Actor.myAnimationControllers to access specific animation triggers. Depending on how severe the error for just flat-out bad calling a trigger is, this may not be necessary