Resynced Ego's Teleport
This was done by splitting Ego's teleport into two animations, an in and an out, that are played back-to-back. This allows stuff (such as the literal teleport) to happen in between.

Currently, Ego will not disappear between anims. This will be resolved by adding blank frame to the in anims or changing Ego's alpha color value in between the animations.