Readjusted dragback values for EgoSonic
It just started working. I'm not really sure why, but my guess is that the ability prefers being wrapped in its WAC since those components are on the same game object, where earlier I was testing it with the user (Ego) not being on the same gameObject as the ability (WeaponAbility).