PlayerAbilityVFX Draft 1
A private class inside of PlayerAbility to handle VFX that occur specifically when the player uses an ability. Currently, it's only programmed to swap the player's material while they have an ability active.

The class hasn't been tested, but it is never instantiated and never referenced, so that's not a problem right now. Appropriate testing will be done once that is the case.