Progress on better Ego2Teleport bounds checking
Actors can use Ego2Movement to teleport out of bounds. For the static version, preventing this is up to the user, but Ego2 has the telemesh to mark valid teleportation locations.

The goal here is to take a provided vector3 and lerp it until it is inside of the telemesh before the teleport starts. Currently non-functional, Ego2 will teleport as normal.

This did involve moving the random-destination teleport into Ego2Movement from Ego2Actor