Adjusted timings for sound effects
Ego2's Swords, Fire, and Teleports now have correct sounds.

The laser's sound is a little bit off due to it not using the same timings as the actual ability. This will be adjusted either with individual sound files for each of the laser stages or by adjusting the laser's timing